+++ src/main/java/dev/doctor4t/arsenal/mixin/ItemMixin.java
@@ -38,7 +38,7 @@ public class ItemMixin {
         }
     }
 
-    @Inject(method = "inventoryTick", at = @At("HEAD"), cancellable = true)
+    @Inject(method = "inventoryTick", at = @At("HEAD"))
     private void arsenal$throw(ItemStack stack, World world, Entity entity, int slot, boolean selected, CallbackInfo ci) {
         if (stack.isOf(Items.TRIDENT) && entity instanceof PlayerEntity player) {
             WeaponOwnerComponent weaponOwnerComponent = ArsenalComponents.WEAPON_OWNER_COMPONENT.get(stack);
@@ -46,5 +46,11 @@ public class ItemMixin {
         }
     }
 
-
+//    @Inject(method = "finishUsing", at = @At("HEAD"))
+//    private void arsenal$removeFromBackslot(ItemStack stack, World world, Entity entity, int slot, boolean selected, CallbackInfo ci) {
+//        if (stack.isOf(Items.TRIDENT) && entity instanceof PlayerEntity player) {
+//            WeaponOwnerComponent weaponOwnerComponent = ArsenalComponents.WEAPON_OWNER_COMPONENT.get(stack);
+//            weaponOwnerComponent.setOwner(player.getUuid());
+//        }
+//    }
 }