+++ src/main/java/dev/doctor4t/arsenal/util/BackSlot.java
@@ -3,8 +3,8 @@ package dev.doctor4t.arsenal.util;
 import net.minecraft.inventory.Inventory;
 import net.minecraft.screen.slot.Slot;
 
-public class WeaponSlot extends Slot {
-    public WeaponSlot(Inventory inventory, int index, int x, int y) {
+public class BackSlot extends Slot {
+    public BackSlot(Inventory inventory, int index, int x, int y) {
         super(inventory, index, x, y);
     }
 }